pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "snapshot-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "boto3>=1.40.6",
  "ipykernel>=6.30.1",
  "jupyter>=1.1.1",
  "org-utils",
  "pandas>=2.3.1",
  "pyairtable>=3.1.1",
  "pyarrow>=21.0.0",
  "python-dotenv>=1.1.1",
  "s3fs>=0.4.2",
  "tqdm>=4.67.1",
]

[project.scripts]
make-snapshot = "snapshot_service.writer:main"

[tool.uv.sources]
org-utils    = { path = "../org_utils", editable = true }

# ---- Tell hatch/packaging that code lives under src/
[tool.hatch.build.targets.wheel]
packages = ["src/snapshot_service"]

